# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistema de Ponto para Usinas Solares
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USINAS
...
  /funcionarios:
    get:
      operationId: listFuncionarios
      tags: [funcionarios]
      summary: List all funcionarios
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: usinaId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of funcionarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Funcionario"
    post:
      operationId: createFuncionario
      tags: [funcionarios]
      summary: Create funcionario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFuncionarioBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funcionario"
  /funcionarios/{id}:
    get:
      operationId: getFuncionario
      tags: [funcionarios]
      summary: Get funcionario by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Funcionario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funcionario"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFuncionario
      tags: [funcionarios]
      summary: Update funcionario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFuncionarioBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funcionario"
    delete:
      operationId: deleteFuncionario
      tags: [funcionarios]
      summary: Delete funcionario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /funcionarios/{id}/historico:
    get:
      operationId: getFuncionarioHistorico
      tags: [funcionarios]
      summary: Get historico de ponto for funcionario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: dataInicio
          in: query
          schema:
            type: string
        - name: dataFim
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Historico
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistroPonto"

  # REGISTROS DE PONTO
...
  /usuarios:
    get:
      operationId: listUsuarios
      tags: [auth]
      summary: List usuarios
      responses:
        "200":
          description: List of usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
    post:
      operationId: createUsuario
      tags: [auth]
      summary: Create usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUsuarioBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
  /usuarios/{id}:
    patch:
      operationId: updateUsuario
      tags: [auth]
      summary: Update usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUsuarioBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
    delete:
      operationId: deleteUsuario
      tags: [auth]
      summary: Delete usuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # AUDITORIA
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/Usuario"
        token:
          type: string
      required:
        - user
        - token
    Usuario:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        perfil:
          type: string
          enum: [super_admin, administrador, supervisor]
        usinaId:
          type: ["integer", "null"]
        ativo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nome
        - email
        - perfil
        - usinaId
        - ativo
        - createdAt
    CreateUsuarioBody:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
        password:
          type: string
        perfil:
          type: string
          enum: [super_admin, administrador, supervisor]
        usinaId:
          type: ["integer", "null"]
      required:
        - nome
        - email
        - password
        - perfil
    UpdateUsuarioBody:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
        perfil:
          type: string
          enum: [super_admin, administrador, supervisor]
        usinaId:
          type: ["integer", "null"]
        ativo:
          type: boolean
        password:
          type: string

    # USINAS
...
    Funcionario:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        cpf:
          type: string
        empresa:
          type: string
        cargo:
          type: string
        usinaId:
          type: ["integer", "null"]
        usinaNome:
          type: ["string", "null"]
        telefone:
          type: ["string", "null"]
        fotoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [ativo, inativo, afastado]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nome
        - cpf
        - empresa
        - cargo
        - usinaId
        - usinaNome
        - telefone
        - fotoUrl
        - status
        - createdAt
    CreateFuncionarioBody:
      type: object
      properties:
        nome:
          type: string
        cpf:
          type: string
        empresa:
          type: string
        cargo:
          type: string
        usinaId:
          type: ["integer", "null"]
        telefone:
          type: ["string", "null"]
        fotoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [ativo, inativo, afastado]
      required:
        - nome
        - cpf
        - empresa
        - cargo
        - status
    UpdateFuncionarioBody:
      type: object
      properties:
        nome:
          type: string
        cpf:
          type: string
        empresa:
          type: string
        cargo:
          type: string
        usinaId:
          type: ["integer", "null"]
        telefone:
          type: ["string", "null"]
        fotoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [ativo, inativo, afastado]

    # REGISTROS
    RegistroPonto:
      type: object
      properties:
        id:
          type: integer
        funcionarioId:
          type: integer
        funcionarioNome:
          type: string
        funcionarioCpf:
          type: string
        usinaId:
          type: integer
        usinaNome:
          type: string
        dataHoraEntrada:
          type: string
          format: date-time
        dataHoraSaida:
          type: ["string", "null"]
          format: date-time
        totalMinutos:
          type: ["integer", "null"]
        ip:
          type: ["string", "null"]
        dispositivo:
          type: ["string", "null"]
        status:
          type: string
          enum: [aberto, finalizado, ajustado]
        observacoes:
          type: ["string", "null"]
        fotoSelfieUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - funcionarioId
        - funcionarioNome
        - funcionarioCpf
        - usinaId
        - usinaNome
        - dataHoraEntrada
        - dataHoraSaida
        - totalMinutos
        - ip
        - dispositivo
        - status
        - observacoes
        - fotoSelfieUrl
        - createdAt
...